feat: Add a LoadedParam struct for the pytket decoder
